+++ .idea/workspace.xml
@@ -254,22 +254,22 @@
       <screen x="0" y="0" width="1218" height="720" />
     </state>
     <state x="390" y="117" key="FileChooserDialogImpl/0.0.1218.720@0.0.1218.720" timestamp="1584922401398" />
-    <state width="1175" height="185" key="GridCell.Tab.0.bottom" timestamp="1584982795732">
+    <state width="1175" height="185" key="GridCell.Tab.0.bottom" timestamp="1585154354371">
       <screen x="0" y="0" width="1218" height="720" />
     </state>
-    <state width="1175" height="185" key="GridCell.Tab.0.bottom/0.0.1218.720@0.0.1218.720" timestamp="1584982795732" />
-    <state width="1175" height="185" key="GridCell.Tab.0.center" timestamp="1584982795731">
+    <state width="1175" height="185" key="GridCell.Tab.0.bottom/0.0.1218.720@0.0.1218.720" timestamp="1585154354371" />
+    <state width="1175" height="185" key="GridCell.Tab.0.center" timestamp="1585154354371">
       <screen x="0" y="0" width="1218" height="720" />
     </state>
-    <state width="1175" height="185" key="GridCell.Tab.0.center/0.0.1218.720@0.0.1218.720" timestamp="1584982795731" />
-    <state width="1175" height="185" key="GridCell.Tab.0.left" timestamp="1584982795731">
+    <state width="1175" height="185" key="GridCell.Tab.0.center/0.0.1218.720@0.0.1218.720" timestamp="1585154354371" />
+    <state width="1175" height="185" key="GridCell.Tab.0.left" timestamp="1585154354371">
       <screen x="0" y="0" width="1218" height="720" />
     </state>
-    <state width="1175" height="185" key="GridCell.Tab.0.left/0.0.1218.720@0.0.1218.720" timestamp="1584982795731" />
-    <state width="1175" height="185" key="GridCell.Tab.0.right" timestamp="1584982795732">
+    <state width="1175" height="185" key="GridCell.Tab.0.left/0.0.1218.720@0.0.1218.720" timestamp="1585154354371" />
+    <state width="1175" height="185" key="GridCell.Tab.0.right" timestamp="1585154354371">
       <screen x="0" y="0" width="1218" height="720" />
     </state>
-    <state width="1175" height="185" key="GridCell.Tab.0.right/0.0.1218.720@0.0.1218.720" timestamp="1584982795732" />
+    <state width="1175" height="185" key="GridCell.Tab.0.right/0.0.1218.720@0.0.1218.720" timestamp="1585154354371" />
     <state width="1175" height="220" key="GridCell.Tab.1.bottom" timestamp="1584922912540">
       <screen x="0" y="0" width="1218" height="720" />
     </state>
