+++ src/ProductOrdering/Main.java
@@ -1,4 +1,4 @@
-package sample;
+package ProductOrdering;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
