--- src/ProductOrdering/database/Record.java
@@ -1,7 +1,0 @@
-package ProductOrdering.database;
-
-import java.sql.SQLException;
-
-public class Record{
-}
-
